Crate toml_parser

TOML lexer and parser

Characteristics:

Error recovery

Lazy validation

forbid(unsafe) by default, requiring the unsafe feature otherwise

no_std support, including putting users in charge of allocation choices (including not allocating)

Full parsing is broken into three phases:

Lexing tokens

Parsing tokens (push parser)

Organizing the physical layout into the logical layout, including decoding keys and values

Modules

decoder

Decode raw TOML values into Rust native types

lexer

Lex TOML tokens

parser

A TOML push parser

Structs

ParseError

Raw

A slice of Source

Source

Data encoded as TOML

Span

Location within the Source

Enums

Expected

Traits

ErrorSink

SourceIndex

A helper trait used for indexing operations on Source