toml_parser::parser

Module document

Constants

UNQUOTED_STRING 🔒

Functions

document 🔒

Parse a TOML Document

eof 🔒

ignore_to_newline 🔒

ignore_to_value_close 🔒

Don’t bother recovering until the matching TokenKind

key 🔒

Parse a TOML key

next_token_if 🔒

on_array_open 🔒

Parse an array

on_comment 🔒

Start EOL from TokenKind::Comment

on_expression_dot 🔒

on_expression_key 🔒

Start an expression from a key compatible token type

on_expression_key_val_sep 🔒

on_inline_table_open 🔒

Parse an inline table

on_missing_expression_key 🔒

on_missing_key 🔒

on_missing_std_table 🔒

on_scalar 🔒

Parse a scalar value

on_table 🔒

Start a table from the open token

opt_dot_keys 🔒

Start a key from the first key compatible token type

opt_whitespace 🔒

Parse whitespace, if present

parse_document

Parse lexed tokens into Events

parse_key

Parse lexed tokens into Events

parse_simple_key

Parse lexed tokens into Events

parse_value

Parse lexed tokens into Events

seek 🔒

simple_key 🔒

Parse a TOML simple key

value 🔒

Parse a value

ws_comment_newline 🔒

Parse EOL decor, if present

Type Aliases

Stream 🔒